Add Memory Loader, Progmem Strings, Basic Audio

+++ Code/EyesFreeClockV3/lib/AT25M01/AT25M01.hpp
@@ -10,12 +10,14 @@ typedef enum AT25M01_BlockProtect  {
     EEPROM_PROTECT_ALL
 } AT25M01_BlockProtect;
 
+// TODO: detect and deal with page boundary crossing
+
 class AT25M01{
 public:
     AT25M01(uint8_t cs);
     void write(const uint8_t *src, const uint32_t dst, uint8_t len);
-    void read(const uint8_t src, uint8_t *dst, uint32_t len);
-    uint8_t read(const uint8_t src);
+    void read(const uint32_t src, uint8_t *dst, uint32_t len);
+    uint8_t read(const uint32_t src);
     void setBlockProtect(AT25M01_BlockProtect wpType);
     bool isReady();
